Add code signature verification
Ensure that workers are performing the same code as the spawner process.
Because this check is performed for each DAG->Worker pair, we ensure
that the entire computation graph for this instance has the same core
function implementation.

Note that this only sanity checks the code function itself, not all of
the package dependencies that might affect code results. It's instead
intended as a light check across mismatched environments that might be
caused by editing code in one place and not updating worker functions
elsewhere.