Core instance spawn logic
Previously we linked a DAGDefinition to a particular instance id, which
meant that we had to recreate the definition object upon every desired
spawn. In this commit we change this to allow the call() method of one
definition generate an unlimited amount of instances. This is fulfilled
through the help of a new helper class `DAGInstance` that creates a
virtual layer on top of the DAGDefinition - where method calls still
happen like they are native to the definition, but we pass along the
instance in place of self in order to track the chained logic that
occurs within that one instance.