Add alternative async executor
Previously we spawned a separate runloop during each async function
execution. This is incompatible with clients that are already running a
runloop for another purpose, like in async tests (with pytest-asyncio).
This commit switches it to run async functions inline and refactor the
overall event loop launcher to the worker entrypoint.